Completed hexagonal inventory slots.
