+++ Vivarium/Assets/Prefabs/Environment/Tiles/bone_tip_two.prefab
@@ -23,7 +23,7 @@ Transform:
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 1790910135746046888}
-  m_LocalRotation: {x: -0, y: -1, z: 0, w: 0.0000018795887}
+  m_LocalRotation: {x: -0, y: -1, z: 0, w: 0.0000020544344}
   m_LocalPosition: {x: 0, y: 0, z: 0}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_Children:
