+++ Vivarium/Assets/Prefabs/Environment/Tiles/pencil_eraser.prefab
@@ -23,7 +23,7 @@ Transform:
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 8327679029957881462}
-  m_LocalRotation: {x: 0, y: -0.000001661032, z: 0, w: -1}
+  m_LocalRotation: {x: 0, y: 0.0000017484547, z: 0, w: 1}
   m_LocalPosition: {x: 0, y: 0, z: 0}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_Children:
