+++ Vivarium/Assets/Prefabs/UI/SelectedItemIndicator.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 557057265c1f2674fb8462fcbb06edb6
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
