+++ Vivarium/Assets/Scripts/Common/Constants.cs
@@ -9,7 +9,7 @@ public static class Constants
     public const float AI_DELAY_BETWEEN_ACTIONS = 0.1f;
     public const float PROJECTILE_SPEED = 25f;
     public const int MAX_CHARACTER_ITEMS = 3;
-    public const int MAX_PLAYER_ITEMS = 20;
+    public const int MAX_PLAYER_ITEMS = 21;
     public const float CAMERA_FOLLOW_SKEW = 8f;
     public const float PROJECTILE_HEIGHT = 0.5f;
     public const float AOE_CALCULATION_OFFSET = 0.01f;
