+++ Vivarium/Assets/Scripts/Editor/InventoryItemsViewInspector.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 636f57465980b4d4e95f5808632b3f7f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
