+++ Vivarium/Assets/Scripts/UI/Inventory.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ba2ca88c5ecfaf24cb9cbfe3ca6da1d2
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
