+++ Vivarium/Assets/Scripts/UI/Inventory/InventoryItemsView.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8603a22d3e7b8984b90fcef3253ff29d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
