+++ Vivarium/Assets/Scripts/UI/Inventory/InventorySlot.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ddd301a1ee75b5849b687b187997a8f1
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
