+++ Vivarium/Assets/Scripts/UI/Inventory/InventoryUIController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 64a8c11ca55767e4fabdfbf66d96bfe0
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
