--- Vivarium/Assets/Scripts/UI/InventoryItemsView.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 5636c892911228f40ad9770a0a1a81fe
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
