--- Vivarium/Assets/Scripts/UI/InventorySlot.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 04c5ddef285c87346bc77af8279d97cd
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
