--- Vivarium/Assets/Scripts/UI/InventoryUIController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: de9d66ce5fa637c41a2ece0afc1bae16
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
