+++ Vivarium/Assets/Textures/UI/CommonUIElements/HexagonSelection.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:95b4968060074da6fc6f1b70c8069080b42863bf399bf02f7feafed1a67bb8b0
+size 5002
