+++ Vivarium/Assets/Visuals/Shaders/logPrefab.prefab
@@ -23,7 +23,7 @@ Transform:
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 4269087875755255206}
-  m_LocalRotation: {x: 0, y: 0.000002972375, z: 0, w: 1}
+  m_LocalRotation: {x: 0, y: -0.0000035843352, z: 0, w: -1}
   m_LocalPosition: {x: -0.0000000016136, y: 0, z: 0}
   m_LocalScale: {x: 1.0541, y: 1.0541, z: 1.0541}
   m_Children:
